Fix dimensions of soap box
The thread was blocking proper insertion of the grill, since the grill
was larger than the inner diameter of the thread.

Now shifted calculation of radii around, so that inner thread is always
bigger than the grill. This also allowed to eliminate the special
capInnerRadius, which was used to ensure that the cap always had enough
inner material.