Trying to improve correctness of grill thickness
I am realizing that math classes are too long ago. I seem to be missing
something obvious here.
The x offset is how far pieces get spaced horizontally to achieve the
correct diagonal thickness.

The results are still not perfect (1.8mm when target is 2mm), but the
closest I got so far. Going to keep that as a better approximation.